undo split into two different parsers for mocha & mochawesome
